Make pulpcore a pytest plugin ... once again!
Required PR: https://github.com/pulp/pulp_file/pull/694

[noissue]
