+++ .github/workflows/scripts/script.sh
@@ -111,7 +111,7 @@ cmd_prefix bash -c "django-admin makemigrations --check --dry-run"
 
 if [[ "$TEST" != "upgrade" ]]; then
   # Run unit tests.
-  cmd_prefix bash -c "PULP_DATABASES__default__USER=postgres pytest -v -r sx --color=yes --pyargs pulpcore.tests.unit"
+  cmd_prefix bash -c "PULP_DATABASES__default__USER=postgres pytest -v -r sx --color=yes -p no:pulpcore --pyargs pulpcore.tests.unit"
 fi
 
 # Run functional tests
@@ -147,6 +147,11 @@ if [[ "$TEST" == "upgrade" ]]; then
 
   echo "Restarting in 60 seconds"
   sleep 60
+
+  # Let's reinstall pulpcore, pulp-cli, and pulp-smash so we have the correct dependencies
+  cd ../pulpcore
+  git checkout -f ci_upgrade_test
+  pip install --upgrade --force-reinstall . ../pulp-cli ../pulp-smash
 
   # CLI commands to display plugin versions and content data
   pulp status
