+++ setup.py
@@ -41,5 +41,6 @@ setup(
             "pulpcore-manager = pulpcore.app.manage:manage",
             "pulpcore-worker = pulpcore.tasking.entrypoint:worker",
         ],
+        "pytest11": ["pulpcore = pulpcore.tests.functional"],
     },
 )
